requirements.txt
openai-whisper
ollama
torch
transformers
torchaudio
fastapi
uvicorn
spleeter